+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
-## Added email-subcribe-redirect-block.php
+## Release 1.4.9
+* Updated email-subscribe-redirect-block.php to add submit button hover text color.
+
+
+## Release 1.4.8
+* Added email-subcribe-redirect-block.php
 
 
 ## Relase 1.4.7